'exluded_binary_data_from_XML_Decreased_xml_size'

+++ Tests/Desktop/IO/Reader/DicomFileReaderTest.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2012-2020 fo-dicom contributors.
+﻿// Copyright (c) 2012-2019 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
 using System.IO;
@@ -22,7 +22,7 @@ namespace Dicom.IO.Reader
             if (File.Exists(filename))
             {
                 byte[] buff = File.ReadAllBytes(filename);
-                using (var stream = new MemoryStream(buff))
+                using (MemoryStream stream = new MemoryStream(buff))
                 {
                     dcm = DicomFile.Open(stream, FileReadOption.ReadAll);
                 }
@@ -31,12 +31,7 @@ namespace Dicom.IO.Reader
             // the file shall be completely read in memory, so writing it should be possible even if the stream has been colsed
             string tmpFile = Path.GetTempFileName();
             dcm.Save(tmpFile);
-            Assert.True(File.Exists(tmpFile));
-            Assert.True(new FileInfo(tmpFile).Length > 0);
-            if (File.Exists(tmpFile))
-            {
-                File.Delete(tmpFile);
-            }
+            if (File.Exists(tmpFile)) File.Delete(tmpFile);
         }
 
 
@@ -48,7 +43,7 @@ namespace Dicom.IO.Reader
             if (File.Exists(filename))
             {
                 byte[] buff = File.ReadAllBytes(filename);
-                using (var stream = new MemoryStream(buff))
+                using (MemoryStream stream = new MemoryStream(buff))
                 {
                     dcm = DicomFile.Open(stream, FileReadOption.ReadLargeOnDemand);
                 }
@@ -57,25 +52,19 @@ namespace Dicom.IO.Reader
             // this will save the image without pixels and generate an ObjectDisposedException
             string tmpFile = Path.GetTempFileName();
             Assert.Throws<DicomIoException>(() => dcm.Save(tmpFile));
-            if (File.Exists(tmpFile))
-            {
-                File.Delete(tmpFile);
-            }
+            if (File.Exists(tmpFile)) File.Delete(tmpFile);
         }
 
 
-        [Theory]
-        [InlineData(@".\Test Data\GH355.dcm")]
-        [InlineData(@".\Test Data\10200904.dcm")]
-        [InlineData(@".\Test Data\CR-MONO1-10-chest")]
-        [InlineData(@".\Test Data\TestPattern_RGB.dcm")]
-        public void ReadLargeFileFromStream_WithoutPixeldata(string filename)
+        [Fact]
+        public void ReadLargeFileFromStream_WithoutPixeldata()
         {
             DicomFile dcm = null;
+            string filename = @".\Test Data\GH355.dcm";
             if (File.Exists(filename))
             {
                 byte[] buff = File.ReadAllBytes(filename);
-                using (var stream = new MemoryStream(buff))
+                using (MemoryStream stream = new MemoryStream(buff))
                 {
                     dcm = DicomFile.Open(stream, FileReadOption.SkipLargeTags);
                 }
@@ -93,7 +82,7 @@ namespace Dicom.IO.Reader
             if (File.Exists(filename))
             {
                 byte[] buff = File.ReadAllBytes(filename);
-                using (var stream = new MemoryStream(buff))
+                using (MemoryStream stream = new MemoryStream(buff))
                 {
                     dcm = DicomFile.Open(stream, FileReadOption.SkipLargeTags);
                 }
